+++ lib/rubocop/cop/katalyst_cops.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require_relative "koi/duplicates_association"
+require_relative "koi/table_asset"
 require_relative "koi/table_link_heading"